requirements-pi.txt
Pillow>=9.0.0
spidev
lgpio
rpi-lgpio
numpy
